+++ libs/modal/src/i18n/en.ts
@@ -1,0 +1,10 @@
+import { I18nLocaleDataPackage } from '@spryker/locale';
+
+export default {
+    name: 'modal',
+    data: {
+        'confirmation-title': 'Confirmation',
+        'confirmation-ok': 'Confirm',
+        'confirmation-cancel': 'Cancel',
+    },
+} as I18nLocaleDataPackage;