+++ libs/table.feature.filters/src/lib/table-filters-feature.component.stories.ts
@@ -1,8 +1,7 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { ANALYZE_FOR_ENTRY_COMPONENTS, importProvidersFrom } from '@angular/core';
+import { importProvidersFrom } from '@angular/core';
 import { provideAnimations } from '@angular/platform-browser/animations';
 import { applicationConfig, Meta, moduleMetadata } from '@storybook/angular';
-import { LayoutFlatHostComponent } from '@orchestrator/layout';
 import { DatasourceModule } from '@spryker/datasource';
 import { TableModule } from '@spryker/table';
 import { TableDummyFilterComponent } from '@spryker/table.feature.filters/testing';
@@ -35,11 +34,6 @@ export default {
                         'mock-data': MockTableDatasourceService,
                     } as any),
                 ),
-                {
-                    provide: ANALYZE_FOR_ENTRY_COMPONENTS,
-                    useValue: [LayoutFlatHostComponent, TableDummyFilterComponent],
-                    multi: true,
-                },
             ],
         }),
         moduleMetadata({