+++ libs/table.feature.settings/src/lib/table-settings-feature.component.stories.ts
@@ -1,7 +1,6 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { ANALYZE_FOR_ENTRY_COMPONENTS, importProvidersFrom } from '@angular/core';
+import { importProvidersFrom } from '@angular/core';
 import { provideAnimations } from '@angular/platform-browser/animations';
-import { LayoutFlatHostComponent } from '@orchestrator/layout';
 import { applicationConfig, Meta, moduleMetadata } from '@storybook/angular';
 import { CheckboxModule } from '@spryker/checkbox';
 import { DatasourceModule } from '@spryker/datasource';
@@ -36,11 +35,6 @@ export default {
                 ),
                 importProvidersFrom(LocaleModule.forRoot({ defaultLocale: EN_LOCALE })),
                 importProvidersFrom(EnLocaleModule),
-                {
-                    provide: ANALYZE_FOR_ENTRY_COMPONENTS,
-                    useValue: [LayoutFlatHostComponent],
-                    multi: true,
-                },
             ],
         }),
         moduleMetadata({