+++ libs/layout/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1-next.0](https://github.com/spryker/ui-components/compare/@spryker/layout@1.1.0...@spryker/layout@1.1.1-next.0) (2024-01-04)
+
+
+### Bug Fixes
+
+* **layout:** adjusted header padding ([d8ac844](https://github.com/spryker/ui-components/commit/d8ac844de552d6c484d17c659189d0dd5c350484))
+
+
+
+
+
 # [1.1.0](https://github.com/spryker/ui-components/compare/@spryker/layout@1.1.0-next.0...@spryker/layout@1.1.0) (2023-12-22)
 
 **Note:** Version bump only for package @spryker/layout