+++ libs/unsaved-changes/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1-next.0](https://github.com/spryker/ui-components/compare/@spryker/unsaved-changes@1.0.0...@spryker/unsaved-changes@1.0.1-next.0) (2024-01-04)
+
+
+### Bug Fixes
+
+* **unsaved-changes:** adjust cancel button DE translation ([049adca](https://github.com/spryker/ui-components/commit/049adca2c82b5a38cfedfeeee1496a4788003634))
+
+
+
+
+
 # [1.0.0](https://github.com/spryker/ui-components/compare/@spryker/unsaved-changes@1.0.0-rc.1...@spryker/unsaved-changes@1.0.0) (2022-12-29)
 
 **Note:** Version bump only for package @spryker/unsaved-changes